feat(cmd): revisions for async prompt
this uses caching for cwd, left prompt, right prompt, tooltip, tip
command, and the tooltip coroutine.

the effect is the prompt is always super responsive, except for the
first time after the current directory changes.

the tooltip coroutine is managed a bit differently, to handle some edge
cases a little better.